Move RepositoryTest to base integration package
